weekly-emailing/requirements.txt
xhtml2pdf
python-dotenv
pandas
boto3
psycopg2-binary
altair
vl-convert-python
pytest
Pillow